pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cheshire-sql"
dynamic = ["version"]
authors = [
    {name = "Cheshire Team", email = "contact@example.com"},
]
description = "Terminal-based SQL visualization tool that transforms query results into beautiful ANSI charts"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]
keywords = ["sql", "visualization", "terminal", "charts", "database", "duckdb", "cli"]
dependencies = [
    "duckdb>=1.0.0",
    "plotext>=5.2.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "pyfiglet>=0.8.0",
    "textual>=0.47.0",
    "rich>=13.0.0",
    "clickhouse-driver>=0.2.6",
    "termgraph",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cheshire"
Repository = "https://github.com/yourusername/cheshire"
Documentation = "https://github.com/yourusername/cheshire#readme"
"Bug Reports" = "https://github.com/yourusername/cheshire/issues"

[project.scripts]
cheshire = "cheshire.main:main"

[tool.setuptools]
packages = ["cheshire"]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "cheshire.__version__"}

[tool.setuptools.package-data]
cheshire = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=cheshire --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true